Fix ausearch handling of exit code on 32 bit systems
git-svn-id: http://svn.fedorahosted.org/svn/audit/trunk@977 03a675c2-f56d-4096-908f-63dba836b7e4

+++ src/ausearch-llist.h
@@ -1,6 +1,6 @@
 /*
 * ausearch-llist.h - Header file for ausearch-llist.c
-* Copyright (c) 2005-2008, 2013 Red Hat Inc., Durham, North Carolina.
+* Copyright (c) 2005-2008, 2013-14 Red Hat Inc., Durham, North Carolina.
 * Copyright (c) 2011 IBM Corp.
 * All Rights Reserved.
 *
@@ -55,7 +55,7 @@ typedef struct
   int arch;             // arch
   int syscall;          // syscall
   uint32_t session_id;  // Login session id
-  int exit;             // Syscall exit code
+  long long exit;       // Syscall exit code
   int exit_is_set;      // Syscall exit code is valid
   char *hostname;       // remote hostname
   slist *filename;      // filename list
